refactor(favorit): move logic favorit add into favorit structure
# Conflicts:
#	itachallenge-user/src/main/java/com/itachallenge/user/controller/UserController.java
#	itachallenge-user/src/main/java/com/itachallenge/user/service/UserServiceImpl.java
